revive erc20 bridge (#1615)
Bridge is in Room12, Scrap Confluence. Click on the bell machine thingy

Todo:
- use cache data for account kamis. i'm guessing something like Party?

Note: removed gate 12 for easier testing. We'd be changing Goals anyway, so no biggie here

+++ packages/client/src/network/chain/ERC20.ts
@@ -18,8 +18,8 @@ export function useBalance(address: Address, token: Address, spender: Address) {
   return {
     ...results,
     balances: {
-      allowance: parseTokenBalance(results.data?.[1]?.result as bigint, 18),
-      balance: parseTokenBalance(results.data?.[0]?.result as bigint, 18),
+      allowance: parseTokenBalance(results.data?.[1]?.result, 18),
+      balance: parseTokenBalance(results.data?.[0]?.result, 18),
     },
   };
 }
